backend/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
gunicorn==23.0.0
python-dotenv==1.0.1
requests==2.32.3
appwrite==10.0.0